% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{build_scenario_data}
\alias{build_scenario_data}
\title{Build scenario data for forecasting}
\usage{
build_scenario_data(
  hunting_quotas_sweden,
  hunting_quotas_finland,
  herring_indices_gulf_bothnia = NULL,
  herring_indices_baltic_proper_gulf_finland = NULL
)
}
\arguments{
\item{hunting_quotas_sweden}{numeric vector of hunting quotas for Sweden}

\item{hunting_quotas_finland}{numeric vector of hunting quotas for Finland}

\item{herring_indices_gulf_bothnia}{numeric vector Herring WAA indices for Gulf of Bothnia}

\item{herring_indices_baltic_proper_gulf_finland}{numeric vector Herring WAA indices for Baltic proper and Gulf of Finland}
}
\value{
list to be passed as scenario data
}
\description{
Build scenario data for forecasting
}
