% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{forecast_plot}
\alias{forecast_plot}
\title{Plot forecast}
\usage{
forecast_plot(forecast, future_years = NULL, variables = NULL)
}
\arguments{
\item{forecast}{forecast draws}

\item{future_years}{years}

\item{variables}{variables to plot}
}
\value{
ggplot object
}
\description{
Plot forecast
}
