% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grey_seal_data}
\alias{grey_seal_data}
\title{Grey seal data}
\format{
List of tibbles
}
\usage{
grey_seal_data
}
\description{
Data from Baltic grey seal observations up to 2024. To be used as an input for \code{fit_ipm()}
}
\keyword{datasets}
