% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ipm.R
\name{fit_ipm}
\alias{fit_ipm}
\title{Fit IPM}
\usage{
fit_ipm(
  data,
  species,
  years,
  prior_spec = NULL,
  method = c("sample", "pathfinder", "laplace", "optimize"),
  init = NULL,
  prior_only = FALSE,
  ...
)
}
\arguments{
\item{data}{list of dataframes}

\item{species}{which species}

\item{years}{years for the state process}

\item{prior_spec}{priors}

\item{method}{Stan method}

\item{init}{Initialisation function or draws passed to Stan}

\item{prior_only}{Should only prior be used?}

\item{...}{passed to Stan method}
}
\value{
fitted model
}
\description{
Fits an IPM for seal observations
}
