% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_ipm}
\alias{forecast_ipm}
\title{Forecast IPM}
\usage{
forecast_ipm(
  fit,
  actual_data = NULL,
  scenario_data = NULL,
  future_years,
  species = fit$species,
  ...
)
}
\arguments{
\item{fit}{A fitted IPM object.}

\item{actual_data}{Raw data}

\item{scenario_data}{List of quotas and optionally herring indices}

\item{future_years}{Consecutive calendar years to forecast.}

\item{species}{Either \code{"grey"} or \code{"ringed"}.}

\item{...}{Passed to \code{generate_quantities()}.}
}
\value{
A CmdStan generated-quantities fit.
}
\description{
Forecasts future years from a fitted IPM. When the original data are
supplied, observations in \code{future_years} are automatically included for
leave-future-out validation.
}
