% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_ipm}
\alias{forecast_ipm}
\title{Forecast IPM}
\usage{
forecast_ipm(fit, future_data, species, prior_spec = NULL, ...)
}
\arguments{
\item{fit}{fit object}

\item{future_data}{future data}

\item{species}{which species}

\item{prior_spec}{prior settings}

\item{...}{unused}
}
\value{
forecast results
}
\description{
Forecasts future years from a fitted IPM
}
